+++ tracking_manager/views/ir_model.xml
@@ -16,20 +16,19 @@
                         <field
                             name="automatic_custom_tracking"
                             string="Automatic configuration"
-                            attrs="{'invisible': [('active_custom_tracking', '=', False)]}"
+                            invisible="not active_custom_tracking"
                         />
                         <field
                             name="automatic_custom_tracking_domain"
-                            string="Domain"
-                            attrs="{
-                                'invisible': [('automatic_custom_tracking', '=', False)],
-                                'required': [('automatic_custom_tracking', '=', True)],
-                            }"
+                            invisible="not automatic_custom_tracking"
+                            required="automatic_custom_tracking"
+                            widget="domain"
+                            options="{'model': 'ir.model.fields', 'in_dialog': True, 'foldable': True}"
                         />
                         <label
                             for="update_custom_tracking"
                             string="Update fields configuration"
-                            attrs="{'invisible': [('automatic_custom_tracking', '=', False)]}"
+                            invisible="not automatic_custom_tracking"
                         />
                         <button
                             name="update_custom_tracking"
@@ -37,7 +36,7 @@
                             icon="fa-refresh"
                             type="object"
                             class="btn-secondary"
-                            attrs="{'invisible': [('automatic_custom_tracking', '=', False)]}"
+                            invisible="not automatic_custom_tracking"
                         />
                     </group>
                 </group>
@@ -45,21 +44,23 @@
 
             <xpath expr="//sheet/group[1]" position="before">
                         <div class="oe_button_box" name="button_box">
-                            <button
+
+                <button
                         name="%(ir_model_fields_action)d"
                         type="action"
                         class="oe_stat_button"
                         icon="fa-server"
-                        attrs="{'invisible': [('active_custom_tracking', '=', False)]}"
+                        invisible="not active_custom_tracking"
                     >
                             <field
                             name="tracked_field_count"
                             widget="statinfo"
                             string="Tracked Fields"
                         />
-                            </button>
-                        </div>
-                </xpath>
+                </button>
+            </div>
+            </xpath>
+
         </field>
     </record>
